bot/pyproject.toml
[project]
name = "lms-telegram-bot"
version = "0.1.0"
description = "Telegram bot for LMS integration"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot>=21.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
